feat: command editing
